Restructure optimization modules
