+++ dhnx/__init__.py
@@ -7,3 +7,6 @@ from . import model
 from . import simulation
 from . import dhn_from_osm
 from . import optimization
+from . import optimization_add_components
+from . import optimization_dhs_nodes
+from . import optimization_oemof_heatpipe
