+++ dhnx/optimization_dhs_nodes.py
@@ -13,7 +13,8 @@ SPDX-License-Identifier: MIT
 
 
 import oemof.solph as solph
-from dhnx.optimization_modules import oemof_heatpipe as oh, add_components as ac
+import dhnx.optimization_oemof_heatpipe as oh
+import dhnx.optimization_add_components as ac
 
 
 def add_nodes_dhs(opti_network, gd, nodes, busd):
